Updated Gradle and Renamed Project Name
I updated the Gradle to the newest version and changed the project name to better reflect its purpose now that more than one person is working on it.
