+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Robotics_10650_2024_2025_Code/RobotInitialize.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.williamcode;
+package org.firstinspires.ftc.teamcode.Robotics_10650_2024_2025_Code;
 
 
 import com.qualcomm.hardware.bosch.BNO055IMU;
